+++ generators/app/templates/src/components/layout/_index.js
@@ -1,7 +1,13 @@
+import { AuthenticatedNav } from './components/authenticated-nav'
+import { AuthenticatedNavProvider } from './services/authenticated-nav'
+
 import router from './router'
 
+export default angular
 export default angular
   .module('<%= appname %>.components.layout', [
     'ui.router'
   ])
   .config(router)
+  .component('authenticatedNav', AuthenticatedNav)
+  .provider('AuthenticatedNav', AuthenticatedNavProvider)